pyproject.toml
[project]
name = "igreja-em-foco-backend"
version = "1.0.0"
description = "Backend API para o sistema Igreja em Foco"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-migrate>=4.0.0",
    "python-dotenv>=1.0.0",
    "authlib>=1.3.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.31.0",
    "flask-login>=0.6.0",
    "flask-jwt-extended>=4.6.0",
    "flask-cors>=4.0.0",
    "waitress>=3.0.0",
    "faker>=35.2.2",
]
